Changes to AImodel struct and POC pipeline
Now it should be easy to integrate new model plugins. Similar structure will be rolled out across other plugins shortly. Implemented a hardcoded pipeline for the POC, still need to have (storage and retrieval) as well as webscraping sorted to fully implement. Pipelines will be a YAML but this current implementation will allow demonstration for now until that is fully completed.